Update, change datatype Budget to Bigint, long

+++ ModelLibrary/DBModels/Parameter.cs
@@ -13,7 +13,7 @@ namespace ModelLibrary.DBModels
 
         public int Id { get; set; }
         public int? ProjectId { get; set; }
-        public int? Budget { get; set; }
+        public long? Budget { get; set; }
         public DateTime? StartDate { get; set; }
         public DateTime? Deadline { get; set; }
         public int? ObjectiveTime { get; set; }
